Assignments
Question1: A function using positional parameter that can take two numbers and compute the addition, multiplication and display the answers. 
Question2: Function that takes a list of items and display each item on a line
Question3: Function that takes a list of shopping items and displays the length of the list.